Moved config module out.
